requirements.txt
torch==2.7.1
transformers==4.52.4
accelerate==1.8.1
datasets==3.6.0
evaluate==0.4.6
lm_eval==0.4.9
peft==0.18.1
hydra-core==1.3.2
python-dotenv==1.2.2
pydantic==2.12.5
loguru==0.7.3
einops==0.8.2
flash_attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu12torch2.7cxx11abiTRUE-cp310-cp310-linux_x86_64.whl
numpy==2.2.6
tqdm==4.67.3
nltk==3.9.3
rouge==1.0.1
math-verify==0.9.0
latex2sympy2_extended==1.11.0
python-Levenshtein==0.27.3
langdetect
immutabledict
